backend/requirements.txt
django
djangorestframework
django-cors-headers
dj-rest-auth
djangorestframework-simplejwt
django-allauth
